Refactoring back the Shed Controller
