+++ Assets/_Root/Scripts/CoreControllers/MainController.cs
@@ -11,6 +11,10 @@ using Tool;
 
 internal class MainController : BaseController
 {
+    private readonly ResourcePath _upgradeHandlersDataSourcePath = new ResourcePath("Configs/Upgrades/UpgradeItemConfigDataSource");
+    private readonly ResourcePath _itemsConfigDataSourcePath = new ResourcePath("Configs/Inventory/ItemConfigDataSource");
+    private readonly ResourcePath _invetoryViewPath = new ResourcePath("Prefabs/UI/InventoryView");
+    
     private readonly Transform _placeForUi;
     private readonly ProfilePlayer _profilePlayer;
 
@@ -64,12 +68,38 @@ internal class MainController : BaseController
                 _carSelectController = new CarSelectController(_placeForUi, _profilePlayer);
                 break;
             case GameState.Shed:
-                _shedController = new ShedController(_placeForUi, _profilePlayer);
+                _shedController = new ShedController(_placeForUi, _profilePlayer, CreateUpgradeHandlersRepository(), CreateInventoryController(_placeForUi));
                 break;
             default:
                 DisposeAllControllers();
                 break;
         }
+    }
+    
+    private InventoryController CreateInventoryController(Transform placeForUi) =>
+        new InventoryController(LoadInventoryView(placeForUi), _profilePlayer.Inventory, CreateInventoryItemRepository());
+    
+    private ItemsRepository CreateInventoryItemRepository()
+    {
+        ItemConfig[] itemConfigs = ContentDataSourceLoader.LoadItemConfigs(_itemsConfigDataSourcePath);
+        var repository = new ItemsRepository(itemConfigs);
+        return repository;
+    }
+    
+    private IUpgradeHandlersRepository CreateUpgradeHandlersRepository()
+    {
+        UpgradeItemConfig[] upgradeConfigs = ContentDataSourceLoader.LoadUpgradeItemConfigs(_upgradeHandlersDataSourcePath);
+        var repository = new UpgradeHandlersRepository(upgradeConfigs);
+        return repository;
+    }
+    
+    private InventoryView LoadInventoryView(Transform placeForUi)
+    {
+        GameObject prefab = ResourcesLoader.LoadPrefab(_invetoryViewPath);
+        GameObject objectView = Object.Instantiate(prefab, placeForUi);
+        AddGameObject(objectView);
+            
+        return objectView.GetComponent<InventoryView>();
     }
     
     private void DisposeAllControllers()
